+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # CHANGELOG
 
+## 0.1.2
+  - Added getter `Money.amountAsString`.
 ## 0.1.1
   - Added `Money.fromString()` constructor.
   - Added relational operators (`<`, `<=`, `>`, `>=`).
   
 ## 0.1.0+1
-  - Fixes README.md
+  - Fixes `README.md`.
   
 ## 0.1.0
   - Initial version.